feat: update-mode support @ routes-gen
note: uses swc/core for parsing ts-file-exports

+++ cli/helpers/generateRoutes/index.ts
@@ -5,22 +5,37 @@ import mustache from "mustache";
 
 import routeFileTemplate from "~/templates/mustache/routes.js";
 import evaluateRoutesForGeneration from "./evaluator.js";
+import { appendFile } from "fs";
+import { logger } from "~/utils/logger.js";
 
-export default async function (specFilePath: string, appPath: string) {
-  await mkdirp(`${appPath}/app/controllers`);
+export default async function (specFilePath: string, appPath: string, isUpdate: boolean) {
+  if (!isUpdate) await mkdirp(`${appPath}/app/controllers`);
 
-  const { parsedSpec, routes } = await evaluateRoutesForGeneration(specFilePath, appPath);
+  const { parsedSpec, routes, existingControllers } = await evaluateRoutesForGeneration(specFilePath, appPath, isUpdate);
 
   await Promise.allSettled(
     parsedSpec?.tags.map(async (tag) => {
+      const controllerExists = Boolean(existingControllers?.[tag.name]);
       const taggedRoutes = routes.filter(({ tag: routeTag }) => routeTag === tag.name);
+
       const mustachePayload = {
         routes: taggedRoutes,
+        isNew: !controllerExists,
         detail: function () {
-          return this.description || this.summary ? `* description: ${this.description || this.summary}` : "";
+          return this.description || this.summary ? `@description ${this.description || this.summary}` : "";
         },
       };
+
       const fileContent = mustache.render(routeFileTemplate, mustachePayload);
+
+      if (controllerExists)
+        return await appendFile(`${appPath}/app/controllers/${tag.name}.ts`, fileContent, (err) => {
+          if (!err) return;
+          logger.error("An Error occurred while appending to file: ", `/controllers/${tag.name}.ts`);
+          logger.info(err);
+          process.exit(1);
+        });
+
       await writeFile(`${appPath}/app/controllers/${tag.name}.ts`, fileContent);
     })
   );
